add transforms to cameras, add vision subsystem that creates 4 cameras
